+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
+- ProfileParameterValidator now resolves named models before field references
 - Buffer serialization in request body
+
 
 ## [1.0.0] - 2021-11-04
 ### Added